task_11

+++ restaurants/views.py
@@ -59,6 +59,8 @@ def restaurant_detail(request, restaurant_id):
     return render(request, 'detail.html', context)
 
 def restaurant_create(request):
+    if request.user.is_anonymous:
+        return redirect("signin")
     form = RestaurantForm()
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES)
@@ -75,6 +77,9 @@ def restaurant_create(request):
 def item_create(request, restaurant_id):
     form = ItemForm()
     restaurant = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff and request.user != restaurant.owner:
+        return redirect("no-access")
+
     if request.method == "POST":
         form = ItemForm(request.POST)
         if form.is_valid():
@@ -90,19 +95,33 @@ def item_create(request, restaurant_id):
 
 def restaurant_update(request, restaurant_id):
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
+    if not request.user.is_staff and request.user != restaurant_obj.owner:
+        return redirect("no-access")
+
     form = RestaurantForm(instance=restaurant_obj)
     if request.method == "POST":
         form = RestaurantForm(request.POST, request.FILES, instance=restaurant_obj)
         if form.is_valid():
             form.save()
-            return redirect('restaurant-list')
+            return redirect('restaurant-list') 
     context = {
         "restaurant_obj": restaurant_obj,
         "form":form,
-    }
-    return render(request, 'update.html', context)
+    } 
+    return render(request, 'update.html', context)    
+
+
+
+
 
 def restaurant_delete(request, restaurant_id):
+    if not request.user.is_staff:
+        return redirect("no-access")
     restaurant_obj = Restaurant.objects.get(id=restaurant_id)
     restaurant_obj.delete()
     return redirect('restaurant-list')
+
+
+
+def no_access(request):
+    return render(request, "no_access.html")
